+++ lib/widget/container_decoration.dart
@@ -33,9 +33,9 @@ class ContainerDecoration {
   ///
   /// See also:
   ///
-  ///  * [radius], the (maximum) size of the ink splash.
+  ///  * [borderRadius], the (maximum) size of the ink splash.
   ///  * [splashColor], the color of the splash.
-  ///  * [highlightColor], the color of the highlight.
+  ///  * [splashColor], the color of the highlight.
   ///  * [InkSplash.splashFactory], which defines the default splash.
   ///  * [InkRipple.splashFactory], which defines a splash that spreads out
   ///    more aggressively than the default.
@@ -47,8 +47,8 @@ class ContainerDecoration {
   /// See also:
   ///
   ///  * [splashFactory], which defines the appearance of the splash.
-  ///  * [radius], the (maximum) size of the ink splash.
-  ///  * [highlightColor], the color of the highlight.
+  ///  * [borderRadius], the (maximum) size of the ink splash.
+  ///  * [splashColor], the color of the highlight.
   final Color? splashColor;
 
   const ContainerDecoration({